Code/send_api/mebc_monitoring/requirements.txt
paho-mqtt==1.6.1
requests==2.31.0
influxdb==5.3.1
python-dotenv==1.0.0
